fix(deps): upgrade to sqlalchemy version 2
Bump sqlalchemy to version 2.0.17

There are quite some changes that happened with v2.
We are currently relying on the "auto-commit" feature which was removed.
For the moment we can use a wrapper class to mimic the auto-commit
behavior allowing us to move to v2.

Ultimately, the (db) session management needs some overhaul though.

Signed-off-by: moson-mo <mo-son@mailbox.org>